Add files via upload
Team selection:
Objective-

To make a cluster which stores locations of the cooperative individuals.

Task-

Make the accompanying factors and capacities -

Make an array(address type) which stores just 16 components.
getTeamPlayers() - To return the whole cluster components.
selectJerseyNumber() - It will take just a single contention of uint type and returns the players address from the cluster made previously. a) The contention passed into the capacity should be more noteworthy than equivalent to nothing and not exactly equivalent to 15. [Hint - Use require]

Make an Escrow
Escrow is the utilization of an outsider, which holds a resource or assets before they are moved starting with one party then onto the next.

Objective-Create an Escrow contract.

Task-

Make the accompanying public state factors -

payer (address type) - To store the location of the individual who is paying to the payee.
payee (address type) - To strore the location of the individual who is getting sum from payee.
thirdParty (address type) - To store the location of the outsider who will deliver the asset to the payee.
sum (uint256) - The sum the payee needs to move.
Make the accompanying capacities -

constructor ( ) - It will take three contentions to introduce the payer, payee and sum. Instate the thirdParty with msg.sender inside the constructor.
store ( ) -
a) It will take store sent by the payer just . In the event that the payer isn't the depositer then it ought to print "Source should be the payer" . [Hint - Use require()]

b) If the store is finished by the depositer then it will check whether the sum sent by the payer is not exactly or equivalent to contract balance or not. On the off chance that it isn't then, at that point, print "Cant send more than escrow sum" .[Hint - Use require()]

discharge() -
a) It will check regardless of whether the sum is equivalent to the agreement balance. In the event that not it ought to print "can't deliver assets before full sum is sent" . [Hint - Use require()]

b) If the equilibrium is equivalent to sum. Then, at that point, it will actually look at regardless of whether the delivery ( ) by thirdParty. On the off chance that not print "just thirdParty can deliver reserves" . [Hint - Use require()]

c) If discharge( ) is called by the thirdParty. Then, at that point, it will move the sum to the payee.

balanceOf() - To return the equilibrium of the agreement.

Connection of two strings
Objective-

To conactenate two strings.

Task-

Make the accompanying capacity -

1)concatenate ( ) - This capacity will take two contentions i.e string 1 and string 2. It will connect string 1 and string 2 and will return the linked string.

Make a wallet
Objective-

You need to make a wallet. With the assistance of which a client can store ether to his wallet ,send ether to another record and check his own ether account balance.

Task-

Make the accompanying factors and capacities -

Make a public payable state variable by the name of proprietor to store the location of the proprietor.

constructor ( ) - It will take one contention to initialise the proprietor of the agreement.

store ( ) - To store ether to contract balance.

sendEther ( ) - It will take two contentions. First the location of the recipient to whom we need to send the sum to. Second the amount(uint) that we need to send.

a) This capacity should be called by the proprietor of the agreement. If not, it ought to print "shipper isn't permitted". [Hint - Use require]

b) If the proprietor of the agreement has called this work then, at that point, move the ether to the collector's location.

balanceOf ( ) - To return the equilibrium of the agreement.

The Fibonacci numbers are the numbers in the accompanying number succession. 0, 1, 1, 2, 3, 5, 8, 13, 21, 34, 55, 89, 144, … … ..

Task-

Make a lie ( ) to ascertain and return the nth fibonacci number.

Input : n = 2 Output : 1

Input : n = 9 Output : 34